move saveOrderItem to helper
