+++ app/pages/sections/order-items/additional-services/order-item/controller.js
@@ -1,9 +1,6 @@
 import { getContext, getErrorContext } from './contextCreator';
-import { logger } from '../../../../../logger';
 import commonManifest from './commonManifest.json';
 import { getSelectedPriceManifest } from '../../../../../helpers/controllers/manifestProvider';
-import { postOrderItem } from '../../../../../helpers/api/ordapi/postOrderItem';
-import { putOrderItem } from '../../../../../helpers/api/ordapi/putOrderItem';
 
 const formatFormData = ({ formData }) => ({
   quantity: formData.quantity
@@ -62,46 +59,4 @@ export const getOrderItemErrorPageContext = (params) => {
   };
 
   return getErrorContext(updatedParams);
-};
-
-export const saveOrderItem = async ({
-  orderId,
-  orderItemId,
-  orderItemType,
-  accessToken,
-  serviceRecipientId,
-  serviceRecipientName,
-  itemId,
-  itemName,
-  selectedPrice,
-  formData,
-}) => {
-  try {
-    const response = orderItemId === 'neworderitem'
-      ? await postOrderItem({
-        accessToken,
-        orderId,
-        orderItemType,
-        serviceRecipientId,
-        serviceRecipientName,
-        itemId,
-        itemName,
-        selectedPrice,
-        formData,
-      })
-      : await putOrderItem({
-        accessToken,
-        orderId,
-        orderItemId,
-        formData,
-      });
-    return response;
-  } catch (err) {
-    if (err.response.status === 400 && err.response.data && err.response.data.errors) {
-      logger.info(`Validation errors returned from the API ${JSON.stringify(err.response.data.errors)}`);
-      return err.response.data;
-    }
-    logger.error(`Error saving order item for ${itemName} and ${serviceRecipientName} for order id: ${orderId}`);
-    throw new Error();
-  }
 };
